baseGamaScene作成
